Fixed: Issue #1 (compile error occurs on 64bit OS X)

+++ trunk/minidom/minidom_http_parser.h
@@ -137,7 +137,7 @@ namespace minidom
 			case CONTENT:
 				{
 					char bufSize[64];
-					sprintf( bufSize, "%d", buf-offset );
+					sprintf( bufSize, "%d", (int)(buf-offset) );
 					strK = "offset";
 					strV = bufSize;
 					elemNode = contentNode->add( convertString(strK), true );
